+++ src/Traits/CanBeReactedTo.php
@@ -67,9 +67,9 @@ trait CanBeReactedTo
       ->select(
         'subject_id',
         'relation',
-        'relation_type',
+        'type',
         DB::raw('COUNT(*) as count')
       )
-      ->groupBy('relation', 'relation_type', 'subject_id');
+      ->groupBy('relation', 'type', 'subject_id');
   }
 }